* fixed SpriteSDL::Blit(), stretch variant -- sorry mips :( * this is the right place for me to drop a note about how i admire mips for coding 99% correct zoom-supporting code ... completely blindly, with improperly functioning SpriteSDL::Blit(), stretch variant  - ps. 99% because x and y are 1 pixel "off" :)  - pps and i didnt test what happens if you climb on a floor * mmzoom can now safely accept decimal values * updated downloader tools to pull 8.41 instead of 8.31 * options button now throws "not yet implemented" * logout button now prints out a small status message "logging out..." - temporarily until we get a messagebox for confirming logout with y/n * options and logout buttons now have correct "minifont" font

+++ automap.h
@@ -65,7 +65,7 @@ public:
     Automap();
     ~Automap();
 
-    void renderSelf(int x, int y, int w, int h, const Position& pos, int zoom=1);
+    void renderSelf(int x, int y, int w, int h, const Position& pos, double zoom=1);
     void setTileColor(int x, int y, int z, uint8_t color,uint8_t speedindex);
 
 private:
